refactor(optimizers): enhance function optimizers and introduce SqrtFunctionOptimizer
Improved the handling of various mathematical functions within the optimizers, including the addition of the SqrtFunctionOptimizer. Enhanced existing optimizers to better manage method invocations and parameter types, ensuring compatibility with numeric helper types.

This refactor aims to streamline the optimization process for mathematical operations and improve code maintainability.

+++ Vectorize/ConstExpr.SourceGenerator/Helpers/SyntaxHelpers.cs
@@ -108,7 +108,7 @@ public static class SyntaxHelpers
 			case byte bb:
 				return SyntaxFactory.CastExpression(
 					SyntaxFactory.PredefinedType(SyntaxFactory.Token(SyntaxKind.ByteKeyword)),
-					SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal((int)bb)));
+					SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(bb)));
 			case int i:
 				return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(i));
 			case uint ui:
